move game logic to dedicated module
